web_src/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2.1
django-import-export==4.3.7
phonenumbers==9.0.5
sqlparse==0.5.3
tablib==3.8.0
